+++ packages/eslint-plugin-dialtone/tests/lib/rules/deprecated-base-color-classes.js
@@ -12,9 +12,10 @@ const RuleTester = require('eslint').RuleTester;
 // ------------------------------------------------------------------------------
 
 const ruleTester = new RuleTester({
-  // eslint-disable-next-line n/no-extraneous-require
-  parser: require.resolve('vue-eslint-parser'),
-  parserOptions: { ecmaVersion: 'latest' },
+  languageOptions: {
+    parser: require('vue-eslint-parser'),
+    parserOptions: { ecmaVersion: 'latest' },
+  },
 });
 
 ruleTester.run('deprecated-base-color-classes', rule, {